+++ Carbon/Serialization/Format.lua
@@ -1,0 +1,18 @@
+--[[
+	Carbon for Lua
+	#class Serialization.Format
+	#inherits OOP.Object
+
+	#description {
+		**Static**
+
+		A class describing a way to serialize data.
+	}
+]]
+
+local Carbon = (...)
+local OOP = Carbon.OOP
+
+local Format = OOP:StaticClass()
+
+return Format